requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
